feat(medium-pass-1): add visualization types for quick assess/medium pass (#6224)
#### Details

Adding new visualization types necessitates changes in many places.
Generally speaking the changes fall into the following camps:
- VisualizationConfigurationFactory should not throw errors when a new
visualization type is used for `getConfiguration`. This requires the
medium pass provider be implemented and as such many changes in this PR
are signature updates for the creation of the factory.
- `getStoreData` needs to be updated as this is another place where
Assessment and QuickAssess will differ. As such, they've been removed
from assessment-builder and moved to the config factory.
- Consequently this also means we must update the visualization store to
contain mediumPass data.
- The requirement filter is modified to amend the VisualizationType on
the configs and now uses a requirement/testStep to VisualizationType map
to do so.
- Another minor change: checking the TestMode to be Adhoc instead of
Assessment as QuickAssess and Assessment will share the same behavior.
Don't feel super strong about this change but it made sense to me to at
least rename these methods where applicable.

##### Motivation

Feature work.

##### Context

Overwriting the VisualizationType in the written Assessment
configurations is not ideal. Ultimately, landed with this because we
should move to an architecture where the assessment objects themselves
do not contain VisualizationType but are also added/amended with some
mapping. The changes made here do not make that work harder in the
future so I went ahead with this for now.

This still uses medium pass in the naming because it was written before
the quick assess name change; should not be an issue to change later.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS