fix: ensure default store data is merged with persisted data during VisualizationStore initialization (#6231)
#### Details

When there is persisted data in IndexedDB for the `VisualizationStore`
and `storeData.tests` is missing the `mediumPass` key, the
`WebConfigurationVisualizationFactory` hits an error. This PR overrides
`generateDefaultState` for `VisualizationStore` (called during
`initialize()`), allowing us to ensure that default state is generated
for test modes not present in `persistedData`. This implementation is
based on `AssessmentStore`, which is the only other store that overrides
this method. It pulls initial data generation logic into a separate
class to facilitate testing.

##### Motivation

This PR addresses an issue seen in Canary where adhoc visualizations are
nonfunctional and the service worker logs the following error:
```Error while processing browser RuntimeOnMessage event:  TypeError: Cannot read properties of undefined (reading 'keyboardInteractionAssessment')```


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS

Co-authored-by: madalynrose <madparker@microsoft.com>