feat(medium-pass-1): created AssessmentFunctionalitySwitcher (#6236)
#### Details

This allows an easy way for us to switch different
assessment/medium-pass affiliated objects while keeping components
agnostic to either scenario while also not having to pass everything
down props.

##### Motivation

Feature work.

##### Context

Alternatives explored were:
1. Modify the deps object somewhere before we use any of the
assessment/medium-pass specific functionality (like in
`details-view-content`). Decided against this since it violates our
expected pattern of keeping the deps object the same and not treating it
as a normal React prop.
2. Start passing in these objects through props from a component before
we use any functionality (details-view-content again). Avoided this
because this feels like a lot of unnecessary bloat to add to each
component along the UI tree but it did not violate any of our
principles.

Note: the usage of the switcher is not included in this PR.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS