feat(medium-pass-1): use getProvider and getNavLinkHandler in deps instead of direct usages of classes (#6240)
#### Details

This is a follow-up PR to #6236. This covers **some** of the changes for
the assessments provider and all of the changes of the navLinkHandler.

##### Motivation

Feature work.

##### Context

The other remaining changes for the assessmentsProvider will come in a
future PR.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS