feat(medium-pass-1): wire up target page for quick assess (#6247)
#### Details

A couple changes needed to occur to make this happen:
1. selectorMapHelper now uses visualization configuration factory
instead of the assessments provider.
2. The visualization state change handler switches between the
appropriate store data depending on the pivot selected. I think we don't
necessarily have to do it this way as we have fewer components that rely
on the assessment store data explicitly but for now this is fine and
we're not doing anything complicated with this data as well.


##### Motivation

Feature work.

##### Context

Did do a bit of smoke testing making sure this would not cause issues
when the feature flag is disabled.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS