fix: Add logic to detect failed injection in edge IE mode (#6183)
#### Details

Add logic to detect failed injections into the target page, retry, and
if failure continues, sends a notification to the user.

##### Motivation

Addresses issue
https://github.com/microsoft/accessibility-insights-web/issues/5802

##### Context

When investigating
https://github.com/microsoft/accessibility-insights-web/issues/5802, it
appears that we're failing because we attempt to inject the content
scripts into the target page, the injection fails, and we never handle
the error, which causes the details view to show that it's 'scanning'
forever. To address this, this PR adds logic to handle an injection
failure and, if it persists, send a notification to the user.

I'm open to suggestions for how to display the error when the injection
fails. Right now I'm sending a notification since it feels minimally
intrusive, but originally I implemented an 'injection failed' view
similar to the ['no content'
view](https://github.com/microsoft/accessibility-insights-web/blob/main/src/DetailsView/components/no-content-available/no-content-available.tsx)
to be displayed in the details view. The problem with this was that it
prevented users from accessing settings/ the left nav bar when the
target page cannot be communicated with. This causes issues for our e2e
tests as well because we grant minimal permissions for most of those
tests, meaning often we can't communicate with the target page because
of these limited permissions and as a result the injection failed path
is triggered. This is fine when we're just sending a notification on
injection failed, but would mean significant changes to e2e tests if we
were to use a more intrusive view.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
[#5802](https://github.com/microsoft/accessibility-insights-web/issues/5802)
- [x] Ran `yarn null:autoadd`
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS