feat(medium-pass-2): add and wire up data going from quick assess to assessment (#6305)
#### Details

Creates a module responsible for sending the quick assess data to the
assessment store.

##### Motivation

Feature work.

##### Context

The flow of this is slightly complicated so here is the gist:

1. User action (activates button to transfer data from quick assess to
assessment).
2. Action(initializeTransfer) is invoked.
3. Data transfer store is updated with state indicating that transfer
has begun.
4. The updated store state triggers actions (transfer data **with**
quick assess data as payload) and (finalizeTransfer) in
`QuickAssessToAssessmentConverter` (created in this PR).
5. Assessment store is updated with aforementioned quick assess data.

This flow is this way to ensure that we adhere to our flux architecture
(and is similar to the flow used for injecting content-scripts into the
target page).

Note: transitioning the view to "Assessment" in the switcher is not in
the context of this PR (I think it makes sense to make the UI
responsible for that).

Below is a gif on my local machine where I use the "Start Over" button
to trigger the messaging I've set up in this PR and the previous PR
#6302 (and these changes were only for demo purposes and are not in the
PR). Additionally, the dialog which prevents overwriting an in-flight
Assessment is also not in this PR.


![quickAssessToAssessment](https://user-images.githubusercontent.com/32555133/210121684-9bd5eb0d-1ab7-4617-ae35-061d82c5b9c6.gif)

#### Pull request checklist

<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS