feat(medium-pass-2): add data transfer view store to house confirm dialog state (#6314)
#### Details

A view store for housing the state of the confirmation dialog for
transferring data from quick assess to assessment. It's a view store
because this leads to simpler components + this dialog can/should exist
per tab and is removed on refresh.
 
##### Motivation

feature work.

##### Context

Is not used in this PR; future PR will implement the dialog.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS