feat(medium-pass-2): add Complete button to last quick assess requirement (#6352)
#### Details

Replace next requirement button with 'Complete' for the last requirement
in quick assess, prompting data migration to assessment.

##### Motivation

Feature work.

##### Context

Changes the configuration to return the JSX.Element necessary instead. I
could see us renaming the property to `getFooterButton` or something
that feels more all-encompassing but decided to leave as is for now.

Also changes made to behavior around transferring so that we immediately
go to Overview and reset the navigation state for assessment upon
transfer; this leads to a better experience (and likely what was
intended).

Gif of what that looks like now:

![completeButton](https://user-images.githubusercontent.com/32555133/212442147-913a74f5-1d75-4191-8fc3-aa7835a6cbc0.gif)


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS