refactor: remove ability to remove start over assessment/quick assess option in component (#6401)
#### Details

Removes the possibility of the "Start over Quick Assess" (and
Assessment) option from being disabled because we do not need the
component to be able to do that. This simplifies tests and the
component.

##### Motivation

Clean-up refactor.

##### Context

This is a follow-up to changes in #6399 

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS