build: manual NOTICE file for web@2.37.x release (#6410)
#### Details

This PR implements a workaround for an issue currently impacting the
NOTICE generation task we use in release builds where it's missing some
components that we depend on with the following warning:

```
##[warning]Components missing from notice, add manually (4):
See: https://docs.opensource.microsoft.com/content/using/required-notice-template.html
Not harvested in ClearlyDefined (4):
npm/npmjs/-/appium-adb/9.10.23
npm/npmjs/@appium/types/0.8.3
npm/npmjs/@appium/support/3.1.3
npm/npmjs/@appium/tsconfig/0.2.3
```

To work around this, this PR checks in a manually updated version of
`NOTICE.html` at `/src/NOTICE.html` and disables the build tasks that
would normally overwrite that with a generated file in release builds.

To generate the manually updated file, I started from the generated
version of the file and added entries for the four missing packages at
the end of the big list in the HTML file, matching the formatting of the
other generated entries and using data from the respective packages'
`package.json` and `LICENSE` files.

##### Motivation

Unblock release of web@2.37.2

##### Context

This is a quick workaround to unblock the release. We'll want to follow
up with a separate PR that reverts this one and replaces it with some
mechanism for either "making that NOTICE generation warning a build
break" and/or "automating a script to do these manual additions"

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS