fix: quick assess automated checks getting started rendering (#6451)
#### Details

Fixes the issue where navigating to getting started in Automated Checks
in Quick Assess would break the application.

##### Motivation

Fixes bug.

##### Context

When the selected test sub view is 'getting-started', the step returned
by 'getStep' is null. We must be mindful of this if we are to add the
ability to start over specific requirements in quick assess.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS