fix: Improve Reflow Behavior of FastPass Count Badges (#6452)
#### Details

We removed the special css for the outcome chip container and tab stops
minimal requirement header.
We changed the width of the count subclass in the outcome chip container
class to max-content. This removed the min-width that was in the outcome
chip formatting, as this was requiring that the chip was always a
certain size and it was not resizing properly to adjust to the size of
the content.

<img width="435" alt="Screenshot 2023-02-22 at 3 48 40 PM"
src="https://user-images.githubusercontent.com/70558054/220755408-18bb6de6-1c3d-430e-aa43-16a198c6fbc1.png">

##### Motivation

Addressing issue #6415 

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This is a straightforward styling fix.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: #6415 
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS

---------

Co-authored-by: Marisa Ashour <marisaashour@microsoft.com>
Co-authored-by: Annika Bassey <annikabassey@microsoft.com>