chore: remove unnecessary axe-config rules (#6449)
#### Details

This PR:

* Updates the new `@accessibility-insights/axe-config` package to not
include an entry for `frame-tested`. We include this rule in AI4Web's
scans, but then post-process the results such that we don't present it
to users as an automated checks failure (we instead present it as a
special warning bar). This change means that a user relying on
axe-config won't include `frame-tested`, which will avoid the appearance
of issues reported by `axe-config` but not by automated checks.

##### Motivation

* Preparation for [using the axe-config output in
service](https://github.com/microsoft/accessibility-insights-service/pull/2402)

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS