feat(automated-checks): add assessment data to element based view model converter (#6476)
#### Details

Add logic to convert assessment store data and card selection store data
to element based view model. This logic will be used for visualizing
automated checks failures with the new assessment automated checks page/
infrastructure.

##### Motivation

Feature work

##### Context

Implementation notes:

- Right now assessment store data is never passed into
`getElementBasedViewModel`, so the new logic is not being exercised in
practice. Eventually we will hook it up similarly to the [issues/ needs
review logic in
SelectorMapHelper](https://github.com/microsoft/accessibility-insights-web/blob/main/src/injected/selector-map-helper.ts#L96-L107),
but this is blocked until
https://github.com/microsoft/accessibility-insights-web/pull/6457 is
merged as we will need to pass in `AssessmentStoreData` and the new
`AssessmentCardSelectionStoreData` that is being added in that PR. As a
result, I expect there may still be some bugs in this logic that will
have to be investigated once we can hook things up end to end.
- Much of the logic for `getElementBasedViewModel` is the same
regardless of the type of data that is passed in, the main difference is
just how that data is structured. As a result, to reduce duplication,
I'm using an adapted template pattern here such that we normalize both
types of data into one universal format that can them be used to
generate the element based view model.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS